[FEAT] 마이페이지 이름 충돌 해결 - mypage와myPage ui/#40
